backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pdf2image
Pillow
supabase
python-dotenv
python-multipart
